+++ sandbox/erich/trunk-ood-242/extensions/platform/windows/oodialog/oodUtilities.cpp
@@ -1499,7 +1499,7 @@ RexxMethod2(RexxObjectPtr, dlgutil_screenSize_cls, OPTIONAL_CSTRING, _flag, OPTI
                 goto done_out;
             }
 
-            POINT point = {pixelX, pixelY};
+            POINT point = {(LONG)pixelX, (LONG)pixelY}; // cast avoids C4838
             mapPixelToDu(context, dlgObj, &point, 1);
             duX = point.x;
             duY = point.y;
